+++ previous-versions/sync-for-expenses-version-1/src/codatsyncexpenses/models/operations/get_mapping_options.py
@@ -19,10 +19,10 @@ class GetMappingOptionsRequest:
 class GetMappingOptionsResponse:
     content_type: str = dataclasses.field()
     r"""HTTP response content type for this operation"""
-    raw_response: requests_http.Response = dataclasses.field()
-    r"""Raw HTTP response; suitable for custom response parsing"""
     status_code: int = dataclasses.field()
     r"""HTTP response status code for this operation"""
+    raw_response: requests_http.Response = dataclasses.field()
+    r"""Raw HTTP response; suitable for custom response parsing"""
     mapping_options: Optional[shared_mappingoptions.MappingOptions] = dataclasses.field(default=None)
     r"""Success"""