+++ previous-versions/sync-for-expenses-version-1/src/codatsyncexpenses/models/operations/get_sync_transaction.py
@@ -23,10 +23,10 @@ class GetSyncTransactionRequest:
 class GetSyncTransactionResponse:
     content_type: str = dataclasses.field()
     r"""HTTP response content type for this operation"""
-    raw_response: requests_http.Response = dataclasses.field()
-    r"""Raw HTTP response; suitable for custom response parsing"""
     status_code: int = dataclasses.field()
     r"""HTTP response status code for this operation"""
+    raw_response: requests_http.Response = dataclasses.field()
+    r"""Raw HTTP response; suitable for custom response parsing"""
     classes: Optional[List[shared_transactionmetadata.TransactionMetadata]] = dataclasses.field(default=None)
     r"""Success"""