+++ previous-versions/sync-for-expenses-version-1/src/codatsyncexpenses/models/operations/list_sync_transactions.py
@@ -25,10 +25,10 @@ class ListSyncTransactionsRequest:
 class ListSyncTransactionsResponse:
     content_type: str = dataclasses.field()
     r"""HTTP response content type for this operation"""
-    raw_response: requests_http.Response = dataclasses.field()
-    r"""Raw HTTP response; suitable for custom response parsing"""
     status_code: int = dataclasses.field()
     r"""HTTP response status code for this operation"""
+    raw_response: requests_http.Response = dataclasses.field()
+    r"""Raw HTTP response; suitable for custom response parsing"""
     transaction_metadata_list: Optional[shared_transactionmetadatalist.TransactionMetadataList] = dataclasses.field(default=None)
     r"""Success"""