+++ previous-versions/sync-for-expenses-version-1/src/codatsyncexpenses/models/operations/list_syncs.py
@@ -19,10 +19,10 @@ class ListSyncsRequest:
 class ListSyncsResponse:
     content_type: str = dataclasses.field()
     r"""HTTP response content type for this operation"""
-    raw_response: requests_http.Response = dataclasses.field()
-    r"""Raw HTTP response; suitable for custom response parsing"""
     status_code: int = dataclasses.field()
     r"""HTTP response status code for this operation"""
+    raw_response: requests_http.Response = dataclasses.field()
+    r"""Raw HTTP response; suitable for custom response parsing"""
     classes: Optional[List[shared_companysyncstatus.CompanySyncStatus]] = dataclasses.field(default=None)
     r"""Success"""