+++ previous-versions/sync-for-expenses-version-1/src/codatsyncexpenses/models/operations/update_expense_dataset.py
@@ -34,10 +34,10 @@ class UpdateExpenseDatasetResponseBody:
 class UpdateExpenseDatasetResponse:
     content_type: str = dataclasses.field()
     r"""HTTP response content type for this operation"""
-    raw_response: requests_http.Response = dataclasses.field()
-    r"""Raw HTTP response; suitable for custom response parsing"""
     status_code: int = dataclasses.field()
     r"""HTTP response status code for this operation"""
+    raw_response: requests_http.Response = dataclasses.field()
+    r"""Raw HTTP response; suitable for custom response parsing"""
     object: Optional[UpdateExpenseDatasetResponseBody] = dataclasses.field(default=None)
     r"""Accepted"""