+++ previous-versions/sync-for-expenses-version-1/src/codatsyncexpenses/models/shared/postsync.py
@@ -10,6 +10,7 @@ from typing import List, Optional
 @dataclass_json(undefined=Undefined.EXCLUDE)
 @dataclasses.dataclass
 class PostSync:
-    dataset_ids: Optional[List[str]] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('datasetIds') }})
+    UNSET='__SPEAKEASY_UNSET__'
+    dataset_ids: Optional[List[str]] = dataclasses.field(default=UNSET, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('datasetIds'), 'exclude': lambda f: f is PostSync.UNSET }})