+++ previous-versions/sync-for-expenses-version-1/src/codatsyncexpenses/models/shared/syncfailedwebhook.py
@@ -22,9 +22,12 @@ class SyncFailedWebhook:
     r"""Unique identifier for your SMB in Codat."""
     data: Optional[SyncFailedWebhookData] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('Data'), 'exclude': lambda f: f is None }})
     message: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('Message'), 'exclude': lambda f: f is None }})
-    r"""A human readable message about the webhook."""
+    r"""A human-readable message about the webhook."""
     rule_id: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('RuleId'), 'exclude': lambda f: f is None }})
-    r"""Unique identifier for the rule."""
+    r"""Unique identifier for the rule.
+
+    Deprecated field: This will be removed in a future release, please migrate away from it as soon as possible.
+    """
     rule_type: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('RuleType'), 'exclude': lambda f: f is None }})
     r"""The type of rule."""