+++ previous-versions/sync-for-expenses-version-1/RELEASES.md
@@ -29,3 +29,13 @@ Based on:
 - [python v0.4.0] previous-versions/sync-for-expenses-version-1
 ### Releases
 - [PyPI v0.4.0] https://pypi.org/project/codat-sync-for-expenses-version-1/0.4.0 - previous-versions/sync-for-expenses-version-1
+
+## 2024-03-20 17:20:59
+### Changes
+Based on:
+- OpenAPI Doc prealpha https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Expenses-v1.yaml
+- Speakeasy CLI 1.214.0 (2.286.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.5.0] previous-versions/sync-for-expenses-version-1
+### Releases
+- [PyPI v0.5.0] https://pypi.org/project/codat-sync-for-expenses-version-1/0.5.0 - previous-versions/sync-for-expenses-version-1